activity api response received

+++ app/src/main/java/com/devansh/rvapp/model/ActivityResponse.kt
@@ -1,11 +1,11 @@
 package com.devansh.rvapp.model
 
 data class ActivityResponse(
-    val accessibility: Int,
+    val accessibility: Double,
     val activity: String,
     val key: String,
     val link: String,
     val participants: Int,
-    val price: Int,
+    val price: Double,
     val type: String
 )
